+++ app/models/task.rb
@@ -1,4 +1,5 @@
 class Task < ActiveRecord::Base
+  include Answerable
   include EventStream::Notifiable
   include NestedQuestionable
   include Commentable
@@ -240,6 +241,13 @@ class Task < ActiveRecord::Base
   # Needed for invitations.
   def invitee_role
     "Override me"
+  end
+
+  # Overrides Answerable.  Since Tasks are STI the client needs to be able to
+  # save an Answer with the expected owner type (Task) rather than the specific
+  # subclass type (ie TahiStandardTasks::ReviewerReportTask)
+  def owner_type_for_answer
+    'Task'
   end
 
   private