+++ config/initializers/subscriptions.rb
@@ -11,7 +11,8 @@ Subscriptions.configure do
 
   add 'paper:submitted', \
       Paper::Submitted::EmailCreator,
-      Paper::Submitted::SnapshotPaper
+      Paper::Submitted::SnapshotPaper,
+      Paper::Submitted::EmailCoauthors
 
   add 'paper:initially_submitted', \
       Paper::Submitted::SnapshotPaper,