+++ db/migrate/20170123164614_make_custom_card_models_paranoid.rb
@@ -1,0 +1,8 @@
+# Add deleted_at to tables
+class MakeCustomCardModelsParanoid < ActiveRecord::Migration
+  def change
+    add_column :answers, :deleted_at, :datetime
+    add_column :cards, :deleted_at, :datetime
+    add_column :card_contents, :deleted_at, :datetime
+  end
+end